ci: classify the 2-GPU FSDP test as long
The 3000-second colocated workload should not run in short, nightly, or image scopes. Move its sole registration label from short to long while preserving the workload and GSM8K evaluation threshold.

+++ tests/e2e/short/test_qwen3_0.6B_fsdp_colocated_2xGPU.py
@@ -7,7 +7,7 @@ import miles.utils.external_utils.command_utils as U
 register_cuda_ci(
     est_time=3000,
     suite="stage-c-2-gpu-h200",
-    labels=["short"],
+    labels=["long"],
 )
 
 MODEL_NAME = "Qwen3-0.6B"
